feat: refactor `tools` module as a package
